Fit learning rate argparse type to float

+++ pretrain.py
@@ -15,7 +15,7 @@ import wandb
 parser = argparse.ArgumentParser(description='My training script.')
 parser.add_argument('--batch_size', type=int, default=8, help='Amount of data that pass simultaneously to model')
 parser.add_argument('--epochs', type=int, default=10, help='Number of epochs to train')
-parser.add_argument('--learning_rate', type=int, default=1e-5, help='Control amount of weight change during optimization')
+parser.add_argument('--learning_rate', type=float, default=1e-5, help='Control amount of weight change during optimization')
 parser.add_argument('--load_model', type=str, help='(Optinal) Folder contain saved model (Model must match --model)')
 parser.add_argument('--model', type=str, help='Model to train: coarse, super_resolution or refinement')
 parser.add_argument('--save_model', type=str,  help='(Optinal) Folder to save model')
